<https://webkit.org/b/120027> AXObjectCache's accessibilityEnhancedUserInterfaceEnabled should be initialized on process creation
Reviewed by Sam Weinig.

Right now we only call AXObjectCache::setEnhancedAccessibility() when AppKit sends
us a notification indicating that the value has changed. We should not have to
rely on that to properly initialize this value. It should be done with the
WebProcessCreationParamters.

* Shared/WebProcessCreationParameters.cpp:
(WebKit::WebProcessCreationParameters::encode):
(WebKit::WebProcessCreationParameters::decode):
* Shared/WebProcessCreationParameters.h:
* UIProcess/mac/WebContextMac.mm:
(WebKit::WebContext::platformInitializeWebProcess):
* WebProcess/mac/WebProcessMac.mm:
(WebKit::WebProcess::platformInitializeWebProcess):



Identifier: 137969@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@154302 268f45cc-cd09-0410-ab3c-d52691b4dbfc