jsc shell should destroy VM as a workaround for LLVM's exit-time destructors https://bugs.webkit.org/show_bug.cgi?id=120921
Reviewed by Oliver Hunt.

LLVM's exit-time destructors will fire when we exit. If there is an on-going
FTL compile at exit, which will happen if the VM that triggered the compile
isn't shut down, then we will crash.

We should get rid of LLVM's exit-time destructors. But before we do that, we
should just do a clean VM shutdown to suppress spurious crashes. This will
help in expanding LLVM coverage for now.

* jsc.cpp:
(jscmain):



Identifier: 138843@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155239 268f45cc-cd09-0410-ab3c-d52691b4dbfc