VectorMover should use std::move https://bugs.webkit.org/show_bug.cgi?id=120959
Reviewed by Geoffrey Garen.

Source/JavaScriptCore:

Work around a bug in GCC by changing the type of the callType bitfield
in CallLinkInfo to be unsigned instead of CallType.

* bytecode/CallLinkInfo.h:

Source/WTF:

This lets the compiler use move constructors when moving data, which can be a performance improvement.
If the vector element type isn't movable it will be copied instead.

* wtf/Vector.h:
(WTF::VectorTypeOperations::move):
(WTF::VectorTypeOperations::moveOverlapping):

Identifier: 138862@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155258 268f45cc-cd09-0410-ab3c-d52691b4dbfc